refactor: fix path for app
