+++ src/app/App.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import SearchInput from '../../components/searchInput';
-import SearchButton from '../../components/searchButton';
-import CharacterData from '../../components/characterData/characterData';
+import SearchInput from '../components/searchInput';
+import SearchButton from '../components/searchButton';
+import CharacterData from '../components/characterData/characterData';
 import './styles/app.css';
-import { SearchResult } from '../../types';
-import PageControl from '../../components/pageControl/pageControl';
-import GlobalContext from '../../components/globalContext';
+import { SearchResult } from '../types';
+import PageControl from '../components/pageControl/pageControl';
+import GlobalContext from '../components/globalContext';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 const App = () => {
